finishing x86 side first added dkm support

+++ include/ark/mem.h
@@ -2,12 +2,12 @@
 #ifndef MEM_H
 #define MEM_H
 
-#include <stddef.h>
-#include <stdint.h>
 
-void *memcpy(void *dest, const void *src, size_t n);
-void *memset(void *dest, int val, size_t n);
-void *memmove(void *dest, const void *src, size_t n);
-int   memcmp(const void *a, const void *b, size_t n);
+#include "ark/types.h"
+
+void *memcpy(void *dest, const void *src, u32 n);
+void *memset(void *dest, int val, u32 n);
+void *memmove(void *dest, const void *src, u32 n);
+int   memcmp(const void *a, const void *b, u32 n);
 
 #endif
